+++ leagueapps-wp.php
@@ -92,6 +92,10 @@ add_action(
 
 add_action( 'init', array( 'LeagueAppsWP\\Presentation\\TeamsBlock', 'register' ) );
 
+if ( is_admin() ) {
+	add_action( 'plugins_loaded', array( 'LeagueAppsWP\\Wp\\Admin', 'register' ) );
+}
+
 if ( defined( 'WP_CLI' ) && WP_CLI ) {
 	require_once LAWP_DIR . 'cli/commands.php';
 }